+++ app/controllers/admin/citations_controller.rb
@@ -1,7 +1,17 @@
 class Admin::CitationsController < AdminController
   def index
+    @query = params[:query]
+
+    citations = (
+      if @query
+        Citation.search(@query)
+      else
+        Citation
+      end
+    )
+
     @citations =
-      Citation.
+      citations.
       order(created_at: :desc).
       paginate(page: params[:page], per_page: 50)
   end