+++ app/models/citation.rb
@@ -46,6 +46,12 @@ class Citation < ApplicationRecord
       or(where(citable: info_request_batch.info_requests))
   end
 
+  def self.search(query)
+    where(<<~SQL, query: query)
+      lower(citations.source_url) LIKE lower('%'||:query||'%')
+    SQL
+  end
+
   def applies_to_batch_request?
     citable.is_a?(InfoRequestBatch)
   end